PJT_TBT/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
black==22.10.0
bleach==5.0.1
boto3==1.26.13
botocore==1.29.13
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
dj-database-url==1.0.0
Django==3.2.13
django-appconf==1.0.5
django-bootstrap5==22.1
django-extensions==3.2.1
django-imagekit==4.1.0
django-mathfilters==1.0.0
django-storages==1.13.1
django-summernote==0.8.20.0
django-widget-tweaks==1.4.12
executing==1.1.1
gunicorn==20.1.0
idna==3.3
ipython==8.5.0
jedi==0.18.1
jmespath==1.0.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
tomli==2.0.1
traitlets==5.4.0
typing_extensions==4.4.0
tzdata==2022.2
urllib3==1.26.10
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
